requirements.txt
pyairtable
requests
python-dotenv